Move example dependencies to devDependencies
Forgive me if I'm misunderstanding the package structure, by my impression is that the `app` directory is not used in the distribution build. This patch moves the dependencies that are only used by `app` into devDependencies. I noticed this issue when I was looking through the lockfile and noticed that multiple versions of styled-components and react were being installed.